requirements.txt
requests==2.25.1
google-cloud-datastore==2.1.6
google-cloud-bigtable==2.3.1
google-cloud-core==1.7.2
